#30 refactor: 예약 관리 페이지에서도 사용할 수 있도록 Calendar 모듈화
